+++ BrowserFront/src/app/components/cars-list/cars-list.component.html
@@ -22,7 +22,6 @@
             </div>
           </div>
           <div class="car-item__features">
-            <!-- Przyklad dodawania cech, np. 'High Performance' -->
             <div class="car-item__feature">
               <span class="feature-icon"></span> {{ car.driveType }} Drive
             </div>
