+++ BrowserFront/src/app/components/dashboard/dashboard.component.css
@@ -1,7 +1,7 @@
 .main-container {
-  max-width: 1300px; /* Maksymalna szerokość kontenera */
-  margin: 0 auto; /* Wyśrodkowanie kontenera */
-  padding: 0 15px; /* Dodatkowe odstępy od krawędzi */
+  max-width: 1300px; 
+  margin: 0 auto;
+  padding: 0 15px;
 }
 
 .search-bar {
@@ -9,26 +9,26 @@
   justify-content: space-between;
   align-items: center;
   margin-top: 20px;
-  gap: 1rem; /* Mniejszy odstęp między elementami */
-  padding: 1.2rem 2rem; /* Dodajemy większy padding dla wygody */
+  gap: 1rem; 
+  padding: 1.2rem 2rem; 
   background-color: #ffffff;
-  border-radius: 12px; /* Zaokrąglone rogi */
-  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1); /* Mocniejszy cień dla efektu głębi */
-  transition: all 0.3s ease; /* Dodanie płynnych animacji */
+  border-radius: 12px; 
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1); 
+  transition: all 0.3s ease; 
   margin-bottom: 2.5rem;
 
   &:hover {
-    box-shadow: 0 10px 20px rgba(0, 123, 255, 0.3); /* Efekt przy najechaniu */
+    box-shadow: 0 10px 20px rgba(0, 123, 255, 0.3); 
   }
 
   .input-with-icon {
     position: relative;
     display: inline-block;
-    width: 100%; /* Wydłużenie pola tekstowego z ikoną */
+    width: 100%; 
   }
 
   .search-input {
-    width: 100%; /* Sprawia, że input wypełnia dostępne miejsce */
+    width: 100%; 
     padding: 0.8rem 2.5rem 0.8rem 1rem;
     border: 2px solid #ddd;
     border-radius: 8px;
@@ -53,7 +53,7 @@
   }
 
   .search-input:focus + .search-icon {
-    color: #0056b3; /* Zmiana koloru ikony przy focus */
+    color: #0056b3; 
   }
 
 
@@ -67,7 +67,7 @@
     border-radius: 8px;
     cursor: pointer;
     transition: all 0.3s ease;
-    width: 160px; /* Wydłużenie przycisku */
+    width: 160px;
     
     &:hover {
       background-color: #0056b3;
@@ -111,12 +111,12 @@
 }
 
 .profile-button .profile-icon {
-  font-size: 24px; /* Ustaw rozmiar ikonki */
+  font-size: 24px; 
 }
 
 .profile-button .fa-spinner {
-  font-size: 24px; /* Rozmiar spinnera */
-  color: #007bff; /* Kolor spinnera */
+  font-size: 24px;
+  color: #007bff; 
 }
 
 .search-input{
