+++ BrowserFront/src/app/components/dashboard/dashboard.component.html
@@ -8,7 +8,6 @@
     
     
     <div class="input-with-icon">
-      <!-- Wyświetlanie ikonki tylko, gdy pole tekstowe jest puste -->
       <i *ngIf="!pickupLocation" class="fa fa-search search-icon"></i>
       <input 
         type="text"
