+++ BrowserFront/src/app/components/login/login.component.css
@@ -1,6 +1,5 @@
-/* Tło dla całej strony */
 body {
-    background: linear-gradient(135deg, #74ebd5 0%, #acb6e5 100%); /* Piękny gradient */
+    background: linear-gradient(135deg, #74ebd5 0%, #acb6e5 100%); 
     display: flex;
     justify-content: center;
     align-items: center;
@@ -8,7 +7,6 @@ body {
     margin: 0;
   }
   
-  /* Stylizacja karty */
   .card {
     border-radius: 30px;
     box-shadow: 0 4px 30px rgba(0, 0, 0, 0.2);
@@ -16,40 +14,35 @@ body {
     border: 1px solid rgba(255, 255, 255, 0.5);
   }
   
-  /* Stylizacja nagłówka */
   .card h2 {
     font-family: 'Arial', sans-serif;
     font-weight: bold;
   }
   
-  /* Efekt po najechaniu na przycisk */
   .btn-primary {
-    background-color: #0d47a1; /* Zmiana koloru przycisku */
+    background-color: #0d47a1; 
     transition: background-color 0.3s ease;
   }
   
   .btn-primary:hover {
-    background-color: #0056b3; /* Ciemniejszy kolor przycisku po najechaniu */
+    background-color: #0056b3; 
   }
   
-  /* Dostosowanie kolorów błędów */
   .alert-danger {
     background-color: rgba(255, 0, 0, 0.1);
     color: #d9534f;
     border: 1px solid #d9534f;
   }
 
-  /* Animacja spinnera */
 .spinner {
   border: 4px solid rgba(255, 255, 255, 0.3);
-  border-top: 4px solid #0d47a1; /* Kolor spinnera */
+  border-top: 4px solid #0d47a1; 
   border-radius: 50%;
   width: 40px;
   height: 40px;
   animation: spin 1s linear infinite;
 }
 
-/* Definicja animacji obracania spinnera */
 @keyframes spin {
   0% {
     transform: rotate(0deg);
