+++ BrowserFront/src/app/components/login/login.component.html
@@ -33,7 +33,6 @@
                 {{ errorMessage }}
               </div>
   
-               <!-- Spinner Loader -->
                <div *ngIf="isLoading" class="d-flex justify-content-center mb-3">
                 <div class="spinner"></div>
               </div>
