+++ BrowserFront/src/app/components/registration/registration.component.css
@@ -6,19 +6,19 @@
     font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
     background-color: #f5f7fb;
     color: #333;
-    height: 100vh; /* Pełna wysokość ekranu */
+    height: 100vh;
     display: flex;
-    justify-content: center; /* Wyśrodkowanie poziome */
-    align-items: center; /* Wyśrodkowanie pionowe */
-    margin: 0; /* Usunięcie domyślnych marginesów */
+    justify-content: center;
+    align-items: center;
+    margin: 0; 
   }
   
   .container {
     display: flex;
-    justify-content: center; /* Wyśrodkowanie poziome */
-    align-items: center; /* Wyśrodkowanie pionowe */
+    justify-content: center;
+    align-items: center; 
     width: 100%;
-    height: 100%; /* Pełna wysokość ekranu */
+    height: 100%; 
     padding: 20px;
   }
   
@@ -27,7 +27,7 @@
     border-radius: 12px;
     padding: 40px;
     width: 100%;
-    max-width: 900px; /* Szerokość formularza nie przekracza 900px */
+    max-width: 900px; 
     box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
   }
   
@@ -123,9 +123,8 @@
     background: #3182ce;
   }
   
-  /* Zmiana: Używamy flexbox do wyświetlenia daty wydania i ważności dowodu obok siebie */
   .form-row .form-group {
-    width: calc(50% - 12px); /* Aby elementy były obok siebie */
+    width: calc(50% - 12px); 
   }
   
   @media (max-width: 768px) {
