+++ BrowserFront/src/app/components/registration/registration.component.html
