--- BrowserFront/src/app/error-login/error-login.component.html
@@ -1,1 +1,0 @@
-<p>error-login works!</p>
