--- BrowserFront/src/app/error-login/error-login.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-error-login',
-  standalone: true,
-  templateUrl: './error-login.component.html',
-  styleUrl: './error-login.component.css',
-})
-export class ErrorLoginComponent {}
