+++ BrowserFront/src/app/models/VehicleDetail.model.ts
@@ -1,0 +1,17 @@
+export class VehicleDetail {
+  carId: number = 0;
+  brand: string = '';
+  model: string = '';
+  serialNo: string = '';
+  vinId: string = '';
+  yearOfProduction: number = 0;
+  rentalFrom: Date = new Date();
+  rentalTo: Date = new Date();
+  prize: number = 0.0;
+  driveType: string = '';
+  transmission: string = '';
+  description: string = '';
+  type: string = '';
+  rate: number = 0.0;
+  localization: string = '';
+}
