+++ BrowserFront/src/app/services/cars.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { environment } from '../../environments/environment';
-import { VehicleDetail } from '../cars-list/cars-list.component';
+import { VehicleDetail } from '../models/VehicleDetail.model';
 
 @Injectable({
   providedIn: 'root',
@@ -23,14 +23,12 @@ export class CarsService {
       );
     }
 
-    // Konwersja dat na ISO stringi
     const params = {
       rentalFrom: rentalFrom.toISOString(),
       rentalTo: rentalTo.toISOString(),
       location: location,
     };
 
-    // Wysyłanie zapytania do API
     return this.http.get<VehicleDetail[]>(`${this.apiUrl}/Search`, { params });
   }
 }
