+++ BrowserFront/src/main.ts
@@ -1,14 +1,12 @@
-import { enableProdMode } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
 import { AppComponent } from './app/app.component';
 import { provideRouter, Routes } from '@angular/router';
-import { LoginComponent } from './app/login/login.component';
-import { DashboardComponent } from './app/dashboard/dashboard.component';
-import { ErrorLoginComponent } from './app/error-login/error-login.component';
-import { RegistrationComponent } from './app/registration/registration.component';
+import { LoginComponent } from './app/components/login/login.component';
+import { DashboardComponent } from './app/components/dashboard/dashboard.component';
+import { RegistrationComponent } from './app/components/registration/registration.component';
 import { provideHttpClient } from '@angular/common/http';
-import { ProfileComponent } from './app/profile/profile.component';
-import { EditProfileComponent } from './app/edit-profile/edit-profile.component';
+import { ProfileComponent } from './app/components/profile/profile.component';
+import { EditProfileComponent } from './app/components/edit-profile/edit-profile.component';
 
 export const routes: Routes = [
   { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
@@ -17,7 +15,6 @@ export const routes: Routes = [
   { path: 'dashboard', component: DashboardComponent },
   { path: 'edit-profile', component: EditProfileComponent },
   { path: 'profile', component: ProfileComponent },
-  { path: 'error', component: ErrorLoginComponent },
   { path: '**', redirectTo: 'dashboard' },
 ];
