+++ BrowserFront/src/styles.css
@@ -1,4 +1,3 @@
-/* You can add global styles to this file, and also import other style files */
 .mouse-hover{
     cursor: pointer;
 }
